Abstract okta auth into some spec helper goodness
