+++ modules/veteran_verification/spec/requests/service_history_request_spec.rb
@@ -5,35 +5,11 @@ require 'rails_helper'
 RSpec.describe 'Service History API endpoint', type: :request, skip_emis: true do
   include SchemaMatchers
 
-  let(:token) { 'token' }
-  let(:jwt) do
-    [{
-      'ver' => 1,
-      'jti' => 'AT.04f_GBSkMkWYbLgG5joGNlApqUthsZnYXhiyPc_5KZ0',
-      'iss' => 'https://example.com/oauth2/default',
-      'aud' => 'api://default',
-      'iat' => Time.current.utc.to_i,
-      'exp' => Time.current.utc.to_i + 3600,
-      'cid' => '0oa1c01m77heEXUZt2p7',
-      'uid' => '00u1zlqhuo3yLa2Xs2p7',
-      'scp' => %w[profile email openid service_history.read],
-      'sub' => 'ae9ff5f4e4b741389904087d94cd19b2'
-    }, {
-      'kid' => '1Z0tNc4Hxs_n7ySgwb6YT8JgWpq0wezqupEg136FZHU',
-      'alg' => 'RS256'
-    }]
-  end
-  let(:auth_header) { { 'Authorization' => "Bearer #{token}" } }
-  let(:user) { create(:openid_user, identity_attrs: build(:user_identity_attrs, :loa3)) }
-
-  before(:each) do
-    allow(JWT).to receive(:decode).and_return(jwt)
-    Session.create(uuid: user.uuid, token: token)
-  end
+  let(:scopes) { %w[profile email openid service_history.read] }
 
   context 'with valid emis responses' do
     it 'should return the current users service history with one episode' do
-      with_okta_configured do
+      with_okta_user(scopes) do |auth_header|
         VCR.use_cassette('emis/get_deployment/valid') do
           VCR.use_cassette('emis/get_military_service_episodes/valid') do
             get '/services/veteran_verification/v0/service_history', nil, auth_header
@@ -46,7 +22,7 @@ RSpec.describe 'Service History API endpoint', type: :request, skip_emis: true d
     end
 
     it 'should return the current users service history with multiple episodes' do
-      with_okta_configured do
+      with_okta_user(scopes) do |auth_header|
         VCR.use_cassette('emis/get_deployment/valid') do
           VCR.use_cassette('emis/get_military_service_episodes/valid_multiple_episodes') do
             get '/services/veteran_verification/v0/service_history', nil, auth_header
@@ -66,7 +42,7 @@ RSpec.describe 'Service History API endpoint', type: :request, skip_emis: true d
     end
 
     it 'should match the errors schema', :aggregate_failures do
-      with_okta_configured do
+      with_okta_user(scopes) do |auth_header|
         get '/services/veteran_verification/v0/service_history', nil, auth_header
       end
