+++ modules/veteran_verification/spec/requests/veteran_status_request_spec.rb
@@ -5,35 +5,11 @@ require 'rails_helper'
 RSpec.describe 'Veteran Status API endpoint', type: :request, skip_emis: true do
   include SchemaMatchers
 
-  let(:token) { 'token' }
-  let(:jwt) do
-    [{
-      'ver' => 1,
-      'jti' => 'AT.04f_GBSkMkWYbLgG5joGNlApqUthsZnYXhiyPc_5KZ0',
-      'iss' => 'https://example.com/oauth2/default',
-      'aud' => 'api://default',
-      'iat' => Time.current.utc.to_i,
-      'exp' => Time.current.utc.to_i + 3600,
-      'cid' => '0oa1c01m77heEXUZt2p7',
-      'uid' => '00u1zlqhuo3yLa2Xs2p7',
-      'scp' => %w[profile email openid veteran_status.read],
-      'sub' => 'ae9ff5f4e4b741389904087d94cd19b2'
-    }, {
-      'kid' => '1Z0tNc4Hxs_n7ySgwb6YT8JgWpq0wezqupEg136FZHU',
-      'alg' => 'RS256'
-    }]
-  end
-  let(:auth_header) { { 'Authorization' => "Bearer #{token}" } }
-  let(:user) { create(:openid_user, identity_attrs: build(:user_identity_attrs, :loa3)) }
-
-  before(:each) do
-    allow(JWT).to receive(:decode).and_return(jwt)
-    Session.create(uuid: user.uuid, token: token)
-  end
+  let(:scopes) { %w[profile email openid veteran_status.read] }
 
   context 'with valid emis responses' do
     it 'should return true if the user is a veteran' do
-      with_okta_configured do
+      with_okta_user(scopes) do |auth_header|
         VCR.use_cassette('emis/get_veteran_status/valid') do
           get '/services/veteran_verification/v0/status', nil, auth_header
           expect(response).to have_http_status(:ok)
@@ -44,7 +20,7 @@ RSpec.describe 'Veteran Status API endpoint', type: :request, skip_emis: true do
     end
 
     it 'should return not_confirmed if the user is not a veteran' do
-      with_okta_configured do
+      with_okta_user(scopes) do |auth_header|
         VCR.use_cassette('emis/get_veteran_status/valid_non_veteran') do
           get '/services/veteran_verification/v0/status', nil, auth_header
           expect(response).to have_http_status(:ok)
@@ -61,7 +37,7 @@ RSpec.describe 'Veteran Status API endpoint', type: :request, skip_emis: true do
     end
 
     it 'should match the errors schema', :aggregate_failures do
-      with_okta_configured do
+      with_okta_user(scopes) do |auth_header|
         get '/services/veteran_verification/v0/status', nil, auth_header
 
         expect(response).to have_http_status(:bad_gateway)
