+++ spec/rails_helper.rb
@@ -50,21 +50,6 @@ def with_settings(settings, temp_values)
   end
 end
 
-def with_okta_configured(&block)
-  with_settings(
-    Settings.oidc,
-    auth_server_metadata_url: 'https://example.com/oauth2/default/.well-known/openid-configuration',
-    issuer: 'https://example.com/oauth2/default',
-    audience: 'api://default',
-    base_api_url: 'https://example.com/',
-    base_api_token: 'token'
-  ) do
-    VCR.use_cassette('okta/metadata') do
-      yield block
-    end
-  end
-end
-
 VCR::MATCH_EVERYTHING = { match_requests_on: %i[method uri headers body] }.freeze
 
 VCR.configure do |c|
